requirement.txt
fastapi
pydantic
google-genai
httpx
uvicorn
databricks-sql-connector
databricks-connect
databricks-sdk
pyarrow
JayDeBeApi
dotenv